feat: sort package.json keys before writing (#5)
I got annoyed by the key order in package.json's written by the cli, so I fixed this "bug" in this PR. I based the key order roughly on the order on [this page](https://docs.npmjs.com/files/package.json). keys that are not part of the `keyOrder` array will be moved to the end.